+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const PI: f32 = std::f64::consts::PI as f32;
-pub const INFINITY: f32 = std::f64::INFINITY as f32;
+pub const PI: f64 = std::f64::consts::PI;
+pub const INFINITY: f64 = std::f64::INFINITY;
