+++ src/random.rs
@@ -25,12 +25,12 @@ pub fn random_in_hemisphere(normal: Vec3) -> Vec3 {
     }
 }
 
-pub fn random_double() -> f32 {
+pub fn random_double() -> f64 {
     let mut rng = rand::thread_rng();
-    rng.gen::<f32>()
+    rng.gen::<f64>()
 }
 
-pub fn random_v_double(n: usize) -> std::vec::Vec<f32> {
+pub fn random_v_double(n: usize) -> std::vec::Vec<f64> {
     let mut v = vec![0.0; n];
     for i in 0..n {
         v[i] = random_double();
@@ -38,7 +38,7 @@ pub fn random_v_double(n: usize) -> std::vec::Vec<f32> {
     v
 }
 
-pub fn random_double_in_limit(min: f32, max: f32) -> f32 {
+pub fn random_double_in_limit(min: f64, max: f64) -> f64 {
     min + (max - min) * random_double()
 }
 
@@ -70,7 +70,7 @@ pub fn random_color() -> Vec3 {
     random_unit_vector()
 }
 
-pub fn random_color_in_limit(min: f32, max: f32) -> Vec3 {
+pub fn random_color_in_limit(min: f64, max: f64) -> Vec3 {
     Vec3::new(
         random_double_in_limit(min, max),
         random_double_in_limit(min, max),
