+++ src/texture.rs
@@ -6,7 +6,7 @@ use super::vec3::*;
 use std::rc::*;
 
 pub trait Texture {
-    fn value(&self, u: f32, v: f32, p: Vec3) -> Vec3;
+    fn value(&self, u: f64, v: f64, p: Vec3) -> Vec3;
 }
 
 pub struct SolidColor {
@@ -14,7 +14,7 @@ pub struct SolidColor {
 }
 
 impl SolidColor {
-    pub fn new(x: f32, y: f32, z: f32) -> SolidColor {
+    pub fn new(x: f64, y: f64, z: f64) -> SolidColor {
         SolidColor {
             color_value: Vec3::new(x, y, z),
         }
@@ -28,7 +28,7 @@ impl SolidColor {
 }
 
 impl Texture for SolidColor {
-    fn value(&self, u: f32, v: f32, p: Vec3) -> Vec3 {
+    fn value(&self, u: f64, v: f64, p: Vec3) -> Vec3 {
         self.color_value
     }
 }
@@ -39,7 +39,7 @@ pub struct CheckerTexture {
 }
 
 impl Texture for CheckerTexture {
-    fn value(&self, u: f32, v: f32, p: Vec3) -> Vec3 {
+    fn value(&self, u: f64, v: f64, p: Vec3) -> Vec3 {
         let sines = (10.0 * p.x()).sin() * (10.0 * p.y()).sin() * (10.0 * p.z()).sin();
         if sines < 0.0 {
             self.odd.value(u, v, p)
@@ -62,12 +62,12 @@ impl NoiseTexture {
 }
 
 impl Texture for NoiseTexture {
-    fn value(&self, u: f32, v: f32, p: Vec3) -> Vec3 {
+    fn value(&self, u: f64, v: f64, p: Vec3) -> Vec3 {
         Vec3::new(1.0, 1.0, 1.0) * self.perlin.noise(p)
     }
 }
 
-pub fn get_sphere_uv(p: Vec3) -> (f32, f32) {
+pub fn get_sphere_uv(p: Vec3) -> (f64, f64) {
     let phi = (p.z()).atan2(p.x());
     let theta = p.y().asin();
     let u = 1.0 - (phi + PI) / (2.0 * PI);
