+++ src/world.rs
@@ -78,9 +78,9 @@ impl World {
             for b in range_x.clone() {
                 let choose_mat = random_double();
                 let center = Vec3::new(
-                    (a as i64 - bias) as f32 + 0.9 * random_double(),
+                    (a as i64 - bias) as f64 + 0.9 * random_double(),
                     0.2,
-                    (b as i64 - bias) as f32 + 0.9 * random_double(),
+                    (b as i64 - bias) as f64 + 0.9 * random_double(),
                 );
                 if (center - Vec3::new(4.0, 0.2, 0.0)).length() > 0.9 {
                     world.v_objects.push(Rc::new(MovingSphere::new(
@@ -122,7 +122,7 @@ impl World {
 }
 
 impl Hittable for World {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
         let mut closest_record = None;
         let mut closest_so_far = t_max;
         for object in &self.v_objects {
@@ -134,7 +134,7 @@ impl Hittable for World {
         return closest_record;
     }
 
-    fn bounding_box(&self, t0: f32, t1: f32) -> Option<AABB> {
+    fn bounding_box(&self, t0: f64, t1: f64) -> Option<AABB> {
         if self.v_objects.is_empty() {
             None
         } else {
